use size_t

+++ json.h
@@ -22,14 +22,14 @@ struct json { void *private[4]; };
 
 struct json_valid {
     bool valid;
-    long pos;
+    size_t pos;
 };
 
 // json_valid returns true if the input is valid json data.
 bool json_valid(const char *json_str);
-bool json_validn(const char *str, long len);
+bool json_validn(const char *str, size_t len);
 struct json_valid json_valid_ex(const char *str, int opts);
-struct json_valid json_validn_ex(const char *str, long len, int opts);
+struct json_valid json_validn_ex(const char *str, size_t len, int opts);
 
 // json_parse parses the input data and returns a json value.
 //
@@ -44,7 +44,7 @@ struct json_valid json_validn_ex(const char *str, long len, int opts);
 // the result, beyond the lifetime of the originating 'json_str', then it will
 // likely result in a memory fault or some other type of undefined corruption.
 struct json json_parse(const char *json_str);
-struct json json_parsen(const char *str, long len);
+struct json json_parsen(const char *str, size_t len);
 
 // json_first returns the json's first child.
 //
@@ -79,7 +79,7 @@ enum json_type json_type(struct json json);
 // retain this data for longer that the data from the original 'json_parse'
 // call then you should copy it first, such as:
 // 
-//    long len = json_raw_length(json);
+//    size_t len = json_raw_length(json);
 //    char *raw = malloc(len+1);
 //    memcpy(raw, json_raw(json), len);
 //    raw[len] = '\0';
@@ -87,7 +87,7 @@ enum json_type json_type(struct json json);
 const char *json_raw(struct json json);
 
 // json_raw_length returns the length of the raw json data.
-long json_raw_length(struct json json);
+size_t json_raw_length(struct json json);
 
 // Compare a json's raw data with a C string.
 //
@@ -97,7 +97,7 @@ long json_raw_length(struct json json);
 // or a terminating null-character in the C string is reached.
 // Returns -1, 0, +1 for less-than, equal-to, greater-than.
 int json_raw_compare(struct json json, const char *str);
-int json_raw_comparen(struct json json, const char *str, long len);
+int json_raw_comparen(struct json json, const char *str, size_t len);
 
 // Compares a json string to the provided C string.
 //
@@ -115,11 +115,11 @@ int json_raw_comparen(struct json json, const char *str, long len);
 // If the JSON type is not a JSON_STRING, then it performs a binary comparison
 // of the raw JSON to the C string.
 int json_string_compare(struct json json, const char *str);
-int json_string_comparen(struct json json, const char *str, long len);
+int json_string_comparen(struct json json, const char *str, size_t len);
 
 // json_string_length returns the number of characters needed to represent the
 // JSON as a UTF-8 string.
-long json_string_length(struct json json);
+size_t json_string_length(struct json json);
 
 // json_string_copy copies a json string into the provided C string buffer.
 //
@@ -129,24 +129,24 @@ long json_string_length(struct json json);
 // occurred, for example:
 //    
 //    char buf[64];
-//    long len = json_string_copy(json, str, sizeof(str));
+//    size_t len = json_string_copy(json, str, sizeof(str));
 //    if (len > sizeof(str)-1) {
 //        // ... copy did not complete ...
 //    }
 //
-long json_string_copy(struct json json, char *str, size_t nbytes);
+size_t json_string_copy(struct json json, char *str, size_t nbytes);
 
 // json_array_get returns the child json element at index. This is to be used
 // on json with the type JSON_ARRAY.
-struct json json_array_get(struct json json, long index);
+struct json json_array_get(struct json json, size_t index);
 
 // json_array_count returns the number of elements in a json array.
-long json_array_count(struct json json);
+size_t json_array_count(struct json json);
 
 // json_object_get returns the json value for its key. This is to be used on
 // json with the type JSON_OBJECT.
 struct json json_object_get(struct json json, const char *key);
-struct json json_object_getn(struct json json, const char *key, long len);
+struct json json_object_getn(struct json json, const char *key, size_t len);
 
 // json_get return finds json at the provide path.
 //
@@ -155,7 +155,7 @@ struct json json_object_getn(struct json json, const char *key, long len);
 //    json_get(doc, "name", "last", 0);
 //
 struct json json_get(const char *json_str, const char *path);
-struct json json_getn(const char *json_str, long len, const char *path);
+struct json json_getn(const char *json_str, size_t len, const char *path);
 
 // json_double returns a json's double value.
 double json_double(struct json json);
@@ -179,8 +179,8 @@ bool json_bool(struct json json);
 // to store the escaped JSON string in the C string buffer.
 // If the returned length is greater than n-1, then only a parital copy
 // occurred. In other words the 
-long json_escape(const char *str,char *escaped, size_t n); 
-long json_escapen(const char *str, long len, char *escaped, size_t n); 
+size_t json_escape(const char *str, char *escaped, size_t n); 
+size_t json_escapen(const char *str, size_t len, char *escaped, size_t n); 
 
 // json_ensure ensures that the json value has been entirely scanned.
 // Only useful for when you need to call json_raw_length on large objects and
